Small Amount Of Progress Messing With Some Code, 8/5/2023
Small Amount Of Progress Messing With Some Code. Was just messing around with the Travel Options and Basic Roads code to try and get something resembling something to show up on the map screen, was able to after a bit of testing, hopefully this means I can get something more substantial soon, now that I have some idea what values I need to change to change pixels on the map-screen, 8/5/2023.