Got Player's Current Map Position Physically Marked, 8/5/2023
Got Player's Current Map Position Physically Marked. So it might not seem like much, but it kind of is in some ways. After a bunch of hours of purely just playing around with stuff I was able to actually get the player's current position/pixel on the world-map marked with an "X", so hopefully I can work on this even deeper that I've done this, will see in time, but good progress even with this seemingly small step, 8/5/2023.