ALOT More Playing Around Today, But Good Progress, 8/6/2023
ALOT More Playing Around Today, But Good Progress. So got more stuff done than expected today, that primarily being marking the location where the player last clicked on the map, as well as drawing a "line" between the player's current location and where they clicked. While this part still needs alot of work, I did ALOT of testing and playing around with the line stuff specifically and am pretty happy with the progress involving it, but still needs to be worked on a bunch, still many bugs with it and other stuff as well obviously. Clicking within the current region map seems to mostly work fine, if you are also currently within that region, but clicking onto other regions sometimes has some problems, especially if your current position is outside that region, but it's not always consistent which cause the issues and from where. Likely something to do with the screen-offset stuff based on the region or whatever. Either way, not sure if I'll exactly plan on using this current method in the end, but it is atleast giving me a better idea on how I might do this all in the end, so yeah, pretty good progress and will see if I feel like getting more done tomorrow on this, but not bad so far I think, 8/6/2023.