Learned A Bunch Of Stuff Messing Around Today, 8/8/2023
Learned A Bunch Of Stuff Messing Around Today. Was able to get alot of (mostly silly) stuff done today. The main ones being learning how to better use the pixelBuffer stuff to "draw" images and patterns. In this case I was able to draw a 15x15 "bootleg" fallout style world-map destination crosshair and it works in-game when click around. Even before that added a phallic with the same "sprite sheet" sort of method, very cool, hope to continue on it further with this added knowledge/experience, 8/8/2023.