Got Travel Path To Move/Animate, 8/9/2023
Got Travel Path To Move/Animate. So this took quite a bit of changing around some stuff, but honestly not as much as I initially expected. I decided to add a new Panel to this window, so the movement/animation part would hopefully not interfere with the other panels. Seems to mostly work so far, but I think some stuff with the scaling is still off, will have to play around with that to get it right, the dots look a bit off now, not sure why exactly, but will see. Very good progress though, not certain what I'll work on next now that I got this part working somewhat, maybe actually changing the player's world position based on where they stopped, will see next time I suppose, 8/9/2023.