Basic But Ugly Buttons, Fast Travel That Works, 8/10/2023
Basic But Ugly Buttons, Fast Travel That Works. So after a bunch of hours of messing about, was able to get some ugly, basic but functional buttons onto the map-screen. With this, also got the fast travel to the current position working at a basic level as well, so not bad progress thus far. Next time I work on this I'll possibly try to get time factored in, as well as a clock/visible counter of some kind on the window as the player is "traveling" around the map, will probably use some of the "Convenient Clock" mod's code for the clock/counter display part, will see, 8/10/2023.