Basic Map Clock Working Now, 8/11/2023
Basic Map Clock Working Now. Got the bare basic implementation of the "clock" that ticks up as you are traveling on the regional map working. Also made it so the change in time gets accounted for now when fast traveling. Next thing to work on I'm still not sure, lots of possible paths to branch off from here now, so will see as times goes on I suppose, but nice progress on this one today, 8/11/2023.