More Granular Map Movement Achieved, 8/14/2023
More Granular Map Movement Achieved. So I used a somewhat different method to get the movement to be more granular and "smooth" on the map. That was basically just converting coordinates directly onto the screen and using those, rather than the DFPosition locations alone. Also did a bunch of other stuff to refine the look of the pixel path and such visually, so a bunch of interesting stuff. But I will need to do alot of work to get this use of screen positions to work better when going between regions and such, also when actually fast traveling and looking at the variables of specific location pixels while traveling and such, alot of stuff will need to be done, but it's alot smoother this way, basically the way I initially wanted it, so that's cool. Next time I work on this, might try to get the fast travel stuff to work again with these new positional values, will see, 8/14/2023.