Finally Figured Out How To Reconvert Exact Pos To DFPos, 8/16/2023
Finally Figured Out How To Reconvert Exact Pos To DFPos. This took way longer than it should have, but I seem to have finally figured out how to do this mathematically, so next time I work on this I'll fix the fast travel thing and a bunch of other stuff, then continue with the features and work after all that is tested and working again, good stuff, but I'll need a break today after all of that head-scratching and frustration in finding the solution finally, 8/16/2023.