Good Progress After Realizing An Incorrect Assumption, 9/7/2023
Good Progress After Realizing An Incorrect Assumption. So Interkarma gave me some really important information that changed entirely how I was trying to approach this, atleast involving the mapping of the locations onto the world-map. Daggerfall's "game-world" is actually 1000x500 in size, NOT 320x160 like I originally assumed. With this I realized I could use this to map the locations simply by iterating over all 500,000 pixels basically, and it seems to have worked on the most basic level so far atleast, so very good progress due to this added info and corrected wrong assumption I had. Next time I work on this, I'm still thinking about that now, but probably trying to work with getting the locations clickable maybe? Or zooming in and out on the map, not sure, will see, 9/7/2023.