Got Basic Zoom Feature Working, 9/8/2023
Got Basic Zoom Feature Working. So it is very basic, and was basically entirely copy/pasted from DFU's own map code for zooming, but so far it seems to work fairly well. Next time I work on this, I might try to polish this zooming more, maybe even add the ability to click and drag the zoomed result around, similar to what you can do in DFU, should be interesting, maybe after that is refined better get multiple magnifications and maybe some dedicated UI buttons for zooming or something, that or actually getting the locations to show their names and such when moused over, will see, 9/8/2023.