Got Basic Zooming In & Out At Various Magnifications To Work, 9/9/2023
Got Basic Zooming In & Out At Various Magnifications To Work. Took alot of playing around, but seem to have gotten a basic zooming in and out system working, with different levels of zoom magnifications. Next time I work on this, might try to add an "Update()" method so I can try to get a draggable zoom window working as well, will see. Also need more polish with the zooming stuff as well so each magnification change feels better to control where you want it to focus in on, etc, 9/9/2023.