Zooming In/Out Improved Also View Scrolling With Shift, 9/12/2023
Zooming In/Out Improved Also View Scrolling With Shift. So did some improvement for the controls on zooming in and out, also made it possible to "scroll" the view around while zoomed in at various magnifications if holding shift and dragging the mouse around. This all possible thanks to using the Update() function to check for these inputs every frame. I'll eventually want the zoom stuff to have buttons and such, but for now it is definitely good enough for development purposes atleast. Next I likely plan on trying to get putting the mouse over the location dots to actually do something, such as probably first getting the location names to display as a tool-tip or something, probably simple like that and go further from there, will see, 9/12/2023.