Hovering Mouse Over Valid Region Now Shows Region Name, 9/29/2023
Hovering Mouse Over Valid Region Now Shows Region Name. Through use of a "Bitmap" or basically just taking an imported texture and putting it's color values into a color array, was able to get this function to work. I made a custom recolored version of that Politic.pak image data, so that the regions would be colored so that only the "Red" color values would be used. So I just set the region-index to what the red value should be essentially, and read the color values of that array based on the position of the mouse on the screen, seems to work perfectly fine thus far. Next time I work on this I'm likely going to try and get to working on those UI buttons that will be very important for the finished product of this world-map GUI in the end. But will likely go through alot of experimenting to see what form these buttons and UI in general will fit the best, but will just have to see, 9/29/2023.