Started Work On Basic UI Buttons Framework, 9/30/2023
Started Work On Basic UI Buttons Framework. Mostly just testing stuff out, but this is likely how I'm going to try and do the UI stuff atleast, the framework part of it atleast, so will definitely change as development continues here, but not bad progress atleast, something to work with, 9/30/2023.