Got Left Panel Buttons To Work Basically, Still WIP, 10/3/2023
Got Left Panel Buttons To Work Basically, Still WIP. So just threw together some of the functional parts of the left-panel buttons and a barebones visual representation for them. Now that I have these there atleast, next time I work on this I'm likely going to try and get the "traveling" part working again. Where clicking the map moves the dot around the map and when you exit it fast-travels you to that pixel location, stuff like that. I had this done already before deciding to make an entirely custom window, so hopefully should not be too difficult to copy/paste most of the code for similar results, will see, 10/3/2023.