Got Functional "Travel" Part Of Map Working Again, 10/5/2023
Got Functional "Travel" Part Of Map Working Again. So the stuff that I had working before I decided to make an entirely custom window, I finally got working again using this new window and setup. Did alot of copy/paste work, but also needed to change a good few things to get it working with this new system, but most of it is probably easier now, since I don't have to be concerned with the previous region scaling and stuff like that, for now atleast. Next time I work on this, I'll likely try to add some buttons to center the map view on the player location and stuff like that, as well as hopefully continue on features that I have planned but did not implemented yet at all, so should be more interesting progress from here, will see, 10/5/2023.