Got Few More Simple Buttons Working, Also Some Cleanup, 10/6/2023
Got Few More Simple Buttons Working, Also Some Cleanup. Just some buttons in a new right panel that allow you to center the viewpoint when zoomed in, not much but handy. Also made a new method to reduce the amount of clutter creating all those buttons individually was doing. Next time I work on this, I'm considering working on that "fog of war" idea I was thinking of, so will see on that, 10/6/2023.