Got Very Basic "Fog of War" Implementation Working, 10/9/2023
Got Very Basic "Fog of War" Implementation Working. Extremely WIP feature, but not a bad start I'd say. Next time I work on this I might try to make a toggle button so the map-view automatically centers on the player position, so it's easier to navigate when the Fog of War is active. Also just more refinement and work on this fog of war feature, etc, 10/9/2023.