Started Work On Location Search Framework, 10/16/2023
Started Work On Location Search Framework. In this case there was no progress on the actual parsing text and searching part (most of that is going to be copy/pasted logic from the base DFU map code.) But instead this was for selecting which region to search within, or atleast the visual part of that "interface." So when in the "Region Selection Mode" putting the cursor over regions will highlight them in red, then the idea will be next time to actually get clicking on them in this state to start the text entering and parsing part, which would then lead to searching for locations within that selection region and such. So next time I work on this I'll try to get that clicking on the region part to work and see where it goes from there, etc, 10/16/2023.