Made Some More Good Progress On Location Search Framework, 10/18/2023
Made Some More Good Progress On Location Search Framework. Still not done with this search function, still need to do some more logic for it, and especially in-game testing and polish, etc. But pretty good progress on it today and maybe I can even finish this feature up mostly tomorrow and move onto the next feature and such, will see, 10/18/2023.