Think I Got The Search Function Mostly Finished For Now, 10/21/2023
Think I Got The Search Function Mostly Finished For Now. So it seems to work pretty well, atleast as much as I expected. Zooms into the location you search if it exists, even puts a marker on that location, so for now I think it works perfectly fine for the time being, so that's cool. Next time I work on this, I'm still not certain, but I'm thinking maybe I should start on some of the basic save-data stuff and see how that turns out, see how large of a txt file is created from saving something or multiple things with like 500,000 entries, should be interesting. But yeah, save-data stuff for the fog of war atleast, likely next time, 10/21/2023.