Made Massive Save-File Space Saving Improvements, 10/23/2023
Made Massive Save-File Space Saving Improvements. So after using a method of "manual compression" of the array data into a compressed string representation, then when loading a save translating/decompressing that back into the same array data. The initial ~6MB mod save-data went all the way down to about 4.2 KB, atleast in this case a nearly 1400 times difference! So I'm very happy with this compression method. Sure it won't work in ever instance, but for this time it is a massive difference and appears to work perfectly fine compressing it down and decompressing it back to the original array values. So for now atleast, that's probably all the save-data I'm going to work with and I'm going to probably try working on some other features again, now that I have that basic framework layed down atleast, which I should hopefully be able to fairly easily add onto when needed. But yeah, next time I'll maybe work on the Line of sight stuff maybe? That being changing the "exploration/perception" radius based on what terrain elevation you are standing on and the elevation values nearby, will see, but great progress today, 10/23/2023.