Map Line of Sight Feature Working At Basic Level Now, 10/27/2023
Map Line of Sight Feature Working At Basic Level Now. Took a few days of head-scratching and some frustration and confusion, but I seemingly have gotten the Line of Sight feature to work probably as best as I could expect all things considered, atleast for now. Even made a showcase of it in a video here: https://youtu.be/faYGbOMLvoA?si=_nQh_yZOF3-DjWFA

Besides that though. Next time I work on this I'll probably start working on getting the travel-time/speed more tuned in and working better, so when moving from one pixel to another the time will be different depending on stuff like the terrain difference and climate and such, or just working better in general in the code and more consistent, will see, 10/27/2023.