Very Minor Change To Line of Sight, 10/28/2023
Very Minor Change To Line of Sight.
So it was bothering me a bit that areas higher than where the player was at were being revealed entirely, rather than showing the first "part" of that higher area, but not the areas behind it as well. Kind of hard to explain in words, but I changed one part of this so that would no longer happen, so the LOS is more blocked now depending on where you are standing and what you are looking towards. So you basically have to climb or go up an elevation to reveal most parts of the area within that elevation. Like for a large elevated area you have to be equal or higher than that to actually reveal the larger sections of it. And your vision is much more obstructed if you are in a pit or ravine. I think it's better this way, but I'll likely make this a setting as well, especially depending on feedback and such, will see, but for now I'm leaving it this way. Now onto the travel time related stuff as I mentioned in the previous commit, 10/28/2023.