More Refinement On Travel Time And Travel Modes, 11/2/2023
More Refinement On Travel Time And Travel Modes. Also a note, some of the code in this commit is purely for testing and should be either removed or commented out before proper release of the mod. Anyway, this commit got a good amount done in terms of refining the "Travel Time" aspect down, as well as various travel types and travel modes, taking into consideration player attributes/skills, the terrain heights, and if in water or on land, etc. Next thing I will work on is a bit more refinement on this, including trying to get Climate, Season, and Weather to be taken into account for the Travel Time part atleast. Also having to get the Weather system and such considered and implemented in general. After that maybe do similar for the line of sight/view range stuff, will see, 11/2/2023.