Implemented Pretty Basic Map Weather Simulation, 11/7/2023
Implemented Pretty Basic Map Weather Simulation. Got more tracking stuff for climate values based on the current location on the map. As well as a pretty basic "weather simulation" for when on the world-map. It needs alot of work honestly, but it does work mostly atleast. One of the main issues, that I don't know if I'll be able to fix atm atleast, is that when fast-traveling the weather changes, but often not the value I'm expected. Likely due to the vanilla DFU weather changing logic, which is likely just time based when it decides to change the current weather, but seems like my attempt at changing this does not work, so yeah. Other than that, I think I'm happy enough with the basic weather simulation stuff. So likely the next thing I'm going to try to work on is refining the view range/radius thing, Line of sight, whatever you want to call it, basically alter this value depending on stuff like the travel mode and current weather and such, will take some work but hopefully will work out, after that I'm not sure, but will see, etc, 11/7/2023.