Very Good Progress On Wandering Encounters Framework, 11/11/2023
Very Good Progress On Wandering Encounters Framework. So I got alot more done today than I expected. Basically got a surprisingly functional start or framework for wandering encounters of sorts. Obviously needs alot of work still, but the fact it is working as well as it is from just one day of work, I'm pretty surprised, but yeah cool stuff. Next time I'm going to try and work on is adding a "Wait/Pass-time" button/function, so you can stay in one spot and let the time pass and let encounters move around and such, should not be too difficult, after that not certain, will have to see, 11/11/2023.