Added Wait/Pass-time Button And Function, 11/12/2023
Added Wait/Pass-time Button And Function. Was always going to add this, but decided to do it now since I'd like to be able to see the random encounters more easily by just standing still and letting time pass, but works pretty well so far. I'll likely eventually allow you to set a certain amount of time that you want to wait, similar to in Arcanum with the little drop-down that lets you wait so many hours, or even "until daytime" "until nighttime" stuff like that. Now with this simple function done, I think I'm probably going to try and implement the basic framework for the "vitals" tracking stuff and base aspects of that system and such, will see, 11/12/2023.