Got Basics of Player Vitals & Tracking Working, 11/15/2023
Got Basics of Player Vitals & Tracking Working. Still needs to be worked on and refined alot, but so far it is working fairly well. Also ontop of this added in a button similar to "Wait" but it is "Rest" instead, while in this "mode" of waiting you recover various vitals rather than lose them as time passes, however your vision is reduced to practically nothing while resting, and other stuff to be added later, so far so good, 11/15/2023.