More Refinements To Player Vitals System, 11/16/2023
More Refinements To Player Vitals System. Made it so you can rest while swimming in water, will possibly change this later based on your race and other possible factors, etc. Made it so resting mode automatically stops when all of your vitals are filled to max. Added a nifty animation to the vital bar's alpha value when that bar's % value goes below a certain set amount. Finally, got health and mana into the equation when resting atleast, and made the values that are regenerated somewhat improved. What I plan on doing tomorrow is working on the "passing out" mechanic when your fatigue goes down all the way to 0 and you are helpless to wait for your fatigue to regen up to a certain point during that time, will see how it goes, but good progress so far, 11/16/2023.