Further Vitals Refinement, Can Now Pass Out And Even Die, 11/17/2023
Further Vitals Refinement, Can Now Pass Out And Even Die. When fatigue goes to 0 when on land you pass out and are forced to wait until you recover up to 15% of you max fatigue in this state, during this you are vulnerable to hostile (not yet implemented) and the elements. Also, if you are swimming and pass out, you will begin to drown and rapidly lose health, if you reach 0 health-points before regaining consciousness you will die, closing the map window, spitting you out to the world and instantly setting your health to 0, which kills you, so works pretty well so far actually, only slightly jank but better than expected. Next thing I'm going to try and work on is a proper function to "wait" for a set amount of time, this way I can use this later on for the time-set wait/rest function, as well as other things, in this case setting a required time to set-up camp when trying to rest initially, hopefully it's not too difficult to get working, etc 11/17/2023.