Continued Refinement & Details To Vitals System, 11/17/2023
Continued Refinement & Details To Vitals System. Added a "camp setup" timer for when trying to rest in a new pixel of the map initially, it's surprisingly minor, but does spend some time. Also added weather, terrain height, and more attributes and skills to have an effect on the drain of fatigue when traveling and resting, etc. Still needs work, but not a bad spot thus far I'd say atleast with this system. Next time I work on this I'm likely going to work more on the wandering encounters again. This time trying to attach enemies and such to the encounters, add logic so they can pursue/chase the player, and even when collided with on the map will cause you to "encounter" and fight them as you would expect, etc. Will see how that all goes next time, 11/17/2023.